requirements.txt
Flask>=3.0,<4.0
opencv-python-headless>=4.8,<5.0
numpy>=1.24,<2.0
Pillow>=10.0,<11.0
PyYAML>=6.0,<7.0
paho-mqtt>=1.6,<2.0
